progress

+++ pose_estimation/pose_data_npz.py
@@ -139,7 +139,7 @@ class PoseDataNPZTorch(torch.utils.data.Dataset):
         meta = scene["meta"][()]
         projection = back_project(scene["depth"] / 1000, meta)
 
-        target_pcd = projection[np.where(scene["label"] == obj_id)] # average = 3000
+        target_pcd = projection[np.where(scene["label"] == obj_id)].astype(np.float32) # average = 3000
         source_pcd = self.sample_source_pcd(obj_id, len(target_pcd)) * meta["scales"][obj_id]
         pose = meta["poses_world"][obj_id]
